+++ swing/framework-ui/src/main/java/is/codion/swing/framework/ui/DefaultEntityApplicationPanelBuilder.java
@@ -392,6 +392,7 @@ final class DefaultEntityApplicationPanelBuilder<M extends SwingEntityApplicatio
 			if (displayFrame) {
 				applicationFrame.setVisible(true);
 			}
+			applicationPanel.requestInitialFocus();
 			if (onApplicationStarted != null) {
 				onApplicationStarted.accept(applicationPanel);
 			}