+++ changelog.md
@@ -33,6 +33,7 @@ Codion Change Log
 - EntityApplicationPanel.displayEntityPanelFrame() and displayEntityPanelDialog() now initialize the focus by activating the panel when opened.
 - EntityPanel.activateEvent() renamed activated().
 - EntityApplicationPanel.initializedEvent() renamed initialized().
+- EntityApplicationPanel.requestInitialFocus() added, called by DefaultEntityApplicationBuilder on application start.
 
 ## 0.18.22
 ### is.codion.common.core